refactor: move make_database_metadata_for_endpoint
Now in denokv_testing so use it in test_kv.
